requirements/dev.txt
# Development dependencies
-r base.txt

# Debugging
django-debug-toolbar==4.2.0
ipython==8.19.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Code quality
flake8==6.1.0
black==23.12.1
isort==5.13.2
pylint==3.0.3

# Documentation
Sphinx==7.2.6
sphinx-rtd-theme==2.0.0

# Local development
django-stubs==4.2.6
types-requests==2.31.0.10
